mv CalculateInterfacesAndNetworksForMultusAnnotationUpdate
As part of the SIG-Network code consolidation effort,
move the said function to `pkg/network/vmispec`.

This move will enable the extraction of the logic responsible
for NIC hotplug / hotunplug in a following commit.
