+++ pkg/virt-controller/network/hotplug.go
@@ -24,34 +24,6 @@ import (
 	"kubevirt.io/kubevirt/pkg/network/vmispec"
 )
 
-func CalculateInterfacesAndNetworksForMultusAnnotationUpdate(vmi *v1.VirtualMachineInstance) ([]v1.Interface, []v1.Network, bool) {
-	vmiNonAbsentSpecIfaces := vmispec.FilterInterfacesSpec(vmi.Spec.Domain.Devices.Interfaces, func(iface v1.Interface) bool {
-		return iface.State != v1.InterfaceStateAbsent
-	})
-	ifacesToHotUnplugExist := len(vmi.Spec.Domain.Devices.Interfaces) > len(vmiNonAbsentSpecIfaces)
-
-	ifacesStatusByName := vmispec.IndexInterfaceStatusByName(vmi.Status.Interfaces, nil)
-	ifacesToAnnotate := vmispec.FilterInterfacesSpec(vmiNonAbsentSpecIfaces, func(iface v1.Interface) bool {
-		_, ifaceInStatus := ifacesStatusByName[iface.Name]
-		sriovIfaceNotPlugged := iface.SRIOV != nil && !ifaceInStatus
-		return !sriovIfaceNotPlugged
-	})
-
-	networksToAnnotate := vmispec.FilterNetworksByInterfaces(vmi.Spec.Networks, ifacesToAnnotate)
-
-	ifacesToHotplug := vmispec.FilterInterfacesSpec(ifacesToAnnotate, func(iface v1.Interface) bool {
-		_, inStatus := ifacesStatusByName[iface.Name]
-		return !inStatus
-	})
-	ifacesToHotplugExist := len(ifacesToHotplug) > 0
-
-	isIfaceChangeRequired := ifacesToHotplugExist || ifacesToHotUnplugExist
-	if !isIfaceChangeRequired {
-		return nil, nil, false
-	}
-	return ifacesToAnnotate, networksToAnnotate, isIfaceChangeRequired
-}
-
 func ApplyDynamicIfaceRequestOnVMI(vm *v1.VirtualMachine, vmi *v1.VirtualMachineInstance, hasOrdinalIfaces bool) *v1.VirtualMachineInstanceSpec {
 	vmiSpecCopy := vmi.Spec.DeepCopy()
 	vmiIndexedInterfaces := vmispec.IndexInterfaceSpecByName(vmiSpecCopy.Domain.Devices.Interfaces)
