+++ pkg/virt-controller/network/hotplug_test.go
@@ -22,9 +22,6 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	networkv1 "github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1"
-
-	k8sv1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	v1 "kubevirt.io/api/core/v1"
@@ -36,128 +33,12 @@ import (
 
 var _ = Describe("Network interface hot{un}plug", func() {
 	const (
-		expectNoChange = false
-		expectToChange = !expectNoChange
-
 		testNetworkName1 = "testnet1"
 		testNetworkName2 = "testnet2"
-		testNetworkName3 = "testnet3"
-		testNetworkName4 = "testnet4"
 
 		ordinal = true
 	)
-	DescribeTable("calculate if changes are required",
 
-		func(vmi *v1.VirtualMachineInstance, pod *k8sv1.Pod, expIfaces []v1.Interface, expNets []v1.Network, expToChange bool) {
-			ifaces, nets, exists := network.CalculateInterfacesAndNetworksForMultusAnnotationUpdate(vmi)
-			Expect(ifaces).To(Equal(expIfaces))
-			Expect(nets).To(Equal(expNets))
-			Expect(exists).To(Equal(expToChange))
-		},
-		Entry("when no interfaces exist, change is not required", libvmi.New(), nil, nil, nil, expectNoChange),
-		Entry("when vmi interfaces match pod multus annotation and status, change is not required",
-			libvmi.New(
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName1}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName1}),
-				withInterfaceStatus(v1.VirtualMachineInstanceNetworkInterface{Name: testNetworkName1}),
-			),
-			&k8sv1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-					networkv1.NetworkStatusAnnot: `[
-						{"interface":"net1", "name":"red-net", "namespace": "default"}
-					]`,
-				}},
-			}, nil, nil, expectNoChange,
-		),
-		Entry("when vmi interfaces have an extra interface which requires hotplug",
-			libvmi.New(
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName1, InterfaceBindingMethod: v1.InterfaceBindingMethod{SRIOV: &v1.InterfaceSRIOV{}}}),
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName2}),
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName3}),
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName4, InterfaceBindingMethod: v1.InterfaceBindingMethod{SRIOV: &v1.InterfaceSRIOV{}}}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName1}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName2}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName3}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName4}),
-				withInterfaceStatus(v1.VirtualMachineInstanceNetworkInterface{Name: testNetworkName1}),
-				withInterfaceStatus(v1.VirtualMachineInstanceNetworkInterface{Name: testNetworkName2}),
-			),
-			&k8sv1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-					networkv1.NetworkStatusAnnot: `[
-						{"interface":"net1", "name":"red-net", "namespace": "default"},
-						{"interface":"net2", "name":"blue-net", "namespace": "default"}
-					]`,
-				}},
-			},
-			[]v1.Interface{{Name: testNetworkName1, InterfaceBindingMethod: v1.InterfaceBindingMethod{SRIOV: &v1.InterfaceSRIOV{}}}, {Name: testNetworkName2}, {Name: testNetworkName3}},
-			[]v1.Network{{Name: testNetworkName1}, {Name: testNetworkName2}, {Name: testNetworkName3}},
-			expectToChange,
-		),
-		Entry("when vmi interfaces have an extra SRIOV interface which requires hotplug, change is not required since SRIOV hotplug to a pod is not supported",
-			libvmi.New(
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName1}),
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName2, InterfaceBindingMethod: v1.InterfaceBindingMethod{SRIOV: &v1.InterfaceSRIOV{}}}),
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName3, InterfaceBindingMethod: v1.InterfaceBindingMethod{SRIOV: &v1.InterfaceSRIOV{}}}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName1}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName2}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName3}),
-				withInterfaceStatus(v1.VirtualMachineInstanceNetworkInterface{Name: testNetworkName1}),
-				withInterfaceStatus(v1.VirtualMachineInstanceNetworkInterface{Name: testNetworkName2}),
-			),
-			&k8sv1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-					networkv1.NetworkStatusAnnot: `[
-						{"interface":"net1", "name":"red-net", "namespace": "default"},
-						{"interface":"net2", "name":"blue-net", "namespace": "default"}
-					]`,
-				}},
-			},
-			nil,
-			nil,
-			expectNoChange,
-		),
-		Entry("when a vmi interface has state set to `absent`, requiring hotunplug",
-			libvmi.New(
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName1}),
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName2, State: v1.InterfaceStateAbsent}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName1}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName2}),
-				withInterfaceStatus(v1.VirtualMachineInstanceNetworkInterface{Name: testNetworkName1}),
-				withInterfaceStatus(v1.VirtualMachineInstanceNetworkInterface{Name: testNetworkName2}),
-			),
-			&k8sv1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-					networkv1.NetworkStatusAnnot: `[
-						{"interface":"pod1", "name":"red-net", "namespace": "default"},
-						{"interface":"pod2", "name":"blue-net", "namespace": "default"}
-					]`,
-				}},
-			},
-			[]v1.Interface{{Name: testNetworkName1}},
-			[]v1.Network{{Name: testNetworkName1}},
-			expectToChange,
-		),
-		Entry("when vmi interfaces have an interface to hotplug and one to hot-unplug, given hashed names",
-			libvmi.New(
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName1, State: v1.InterfaceStateAbsent}),
-				libvmi.WithInterface(v1.Interface{Name: testNetworkName2}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName1}),
-				libvmi.WithNetwork(&v1.Network{Name: testNetworkName2}),
-				withInterfaceStatus(v1.VirtualMachineInstanceNetworkInterface{Name: testNetworkName1}),
-			),
-			&k8sv1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-					networkv1.NetworkStatusAnnot: `[
-						{"interface":"pod1a2b3c", "name":"red-net", "namespace": "default"}
-					]`,
-				}},
-			},
-			[]v1.Interface{{Name: testNetworkName2}},
-			[]v1.Network{{Name: testNetworkName2}},
-			expectToChange,
-		),
-	)
 	DescribeTable("apply dynamic interface request on VMI",
 		func(vmiForVM, currentVMI, expectedVMI *v1.VirtualMachineInstance, hasOrdinalIfaces bool) {
 			vm := virtualMachineFromVMI(currentVMI.Name, vmiForVM)
@@ -325,14 +206,6 @@ func bridgeAbsentInterface(name string) v1.Interface {
 	iface := bridgeInterface(name)
 	iface.State = v1.InterfaceStateAbsent
 	return iface
-}
-
-func withInterfaceStatus(ifaceStatus v1.VirtualMachineInstanceNetworkInterface) libvmi.Option {
-	return func(vmi *v1.VirtualMachineInstance) {
-		vmi.Status.Interfaces = append(
-			vmi.Status.Interfaces, ifaceStatus,
-		)
-	}
 }
 
 func virtualMachineFromVMI(name string, vmi *v1.VirtualMachineInstance) *v1.VirtualMachine {
