+++ pkg/virt-controller/watch/vmi/BUILD.bazel
@@ -11,13 +11,13 @@ go_library(
         "//pkg/network/admitter:go_default_library",
         "//pkg/network/multus:go_default_library",
         "//pkg/network/namescheme:go_default_library",
+        "//pkg/network/vmispec:go_default_library",
         "//pkg/storage/backend-storage:go_default_library",
         "//pkg/storage/types:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/hardware:go_default_library",
         "//pkg/util/trace:go_default_library",
         "//pkg/virt-config:go_default_library",
-        "//pkg/virt-controller/network:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/common:go_default_library",
         "//pkg/virt-controller/watch/descheduler:go_default_library",
